+++ CHANGELOG.md
@@ -16,6 +16,8 @@ UNRELEASED
 * Added: polarisations to mapping
 * Fixed: use of REACT_APP_USE_LOCAL_DATA for local development
 * Changed: to show notification when there is an senscalc error after mapping target-observation
+* Changed: Added validation to target fields 
+
 
 8.0.0
 ----------